tests for extract pdf, refactor load pdf
